Move utility functions/classes to new package.
* timestamp related ones -> `util.timestamp`
* NSI related ones (STP) -> `util.nsi`
