+++ tests/util/test_nsi.py
@@ -10,9 +10,10 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+
 import pytest
 
-from supa.utils import parse_stp
+from supa.util.nsi import parse_stp
 
 
 def test_parse_stp_with_year_unqualified() -> None:
